Add Layher SERVICE|DB CONTROLLER|SERVICE

+++ src/main/java/com/example/eutaxratesapi/core/rates/TaxRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TaxRepository extends JpaRepository<Rate,Integer> {
+public interface TaxRepository extends JpaRepository<Rate,Long> {
      Rate findByCountryShortName(String shortName);
 
 }
